+++ api/src/utils/debug.js
@@ -1,8 +1,8 @@
 const debug = require('debug');
-const { app } = require('../parameters').parameters;
 
-if (app.isInEnv('test')) {
-  debug.disable();
-}
-
-module.exports = ns => debug(ns);
+module.exports = ({ parameters }) => (ns) => {
+  if (parameters.app.isInEnv('test')) {
+    debug.disable();
+  }
+  return debug(ns);
+};
