+++ api/tests/utils/tokens.js
@@ -1,6 +1,6 @@
-const AuthenticationService = require('../../src/components/authentication/AuthenticationService');
+const container = require('../../src/container');
 
-const authenticationService = new AuthenticationService();
+const authenticationService = container.resolve('authenticationService');
 // { uid: '123456', iat: 1530490533, exp: 1530497733, iss: 'accounts.google.com' }
 // exp = iat + 2h
 // secret = process.env.ACCESS_TOKEN_SECRET_KEY
